requirements.txt
flask
RPi.GPIO